requirements.txt
Django==4.2
psycopg2-binary==2.9.6
django-allauth==0.54.0
django-filter==23.2
djangorestframework==3.14.0
drf-yasg==1.21.5
gunicorn==20.1.0
whitenoise==6.4.0
redis==4.5.5
celery==5.2.7